fix: Circular dependency in admin panel
