+++ components/admin-panel/queries.ts
@@ -1,0 +1,65 @@
+import { gql } from '@apollo/client';
+
+import { collectiveNavbarFieldsFragment } from '../collective-page/graphql/fragments';
+
+export const adminPanelQuery = gql`
+  query AdminPanel($slug: String!) {
+    account(slug: $slug) {
+      id
+      legacyId
+      slug
+      name
+      isHost
+      type
+      settings
+      isArchived
+      isIncognito
+      imageUrl(height: 256)
+      features {
+        id
+        ...NavbarFields
+        VIRTUAL_CARDS
+        USE_PAYMENT_METHODS
+        EMIT_GIFT_CARDS
+      }
+      policies {
+        REQUIRE_2FA_FOR_ADMINS
+      }
+      ... on AccountWithParent {
+        parent {
+          id
+          slug
+          policies {
+            REQUIRE_2FA_FOR_ADMINS
+          }
+        }
+      }
+      ... on AccountWithHost {
+        hostFeePercent
+        host {
+          id
+          slug
+          name
+          settings
+          policies {
+            EXPENSE_AUTHOR_CANNOT_APPROVE {
+              enabled
+              amountInCents
+              appliesToHostedCollectives
+              appliesToSingleAdminCollectives
+            }
+            COLLECTIVE_MINIMUM_ADMINS {
+              numberOfAdmins
+              applies
+              freeze
+            }
+          }
+        }
+      }
+      ... on AccountWithHost {
+        isApproved
+      }
+    }
+  }
+  ${collectiveNavbarFieldsFragment}
+`;
