+++ components/edit-collective/actions/Archive.js
@@ -7,7 +7,7 @@ import { CollectiveType } from '../../../lib/constants/collectives';
 import { getErrorFromGraphqlException } from '../../../lib/errors';
 import { API_V2_CONTEXT, gqlV1 } from '../../../lib/graphql/helpers';
 
-import { adminPanelQuery } from '../../../pages/admin-panel';
+import { adminPanelQuery } from '../../admin-panel/queries';
 import Container from '../../Container';
 import { getI18nLink } from '../../I18nFormatters';
 import MessageBox from '../../MessageBox';
